+++ src/main/java/ca/on/oicr/gsi/dimsum/controller/rest/DownloadRestController.java
@@ -1,16 +1,13 @@
 package ca.on.oicr.gsi.dimsum.controller.rest;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Map;
 import javax.servlet.http.HttpServletResponse;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,7 +16,9 @@ import org.springframework.web.bind.annotation.RestController;
 import ca.on.oicr.gsi.dimsum.controller.BadRequestException;
 import ca.on.oicr.gsi.dimsum.service.CaseService;
 import ca.on.oicr.gsi.dimsum.util.reporting.Report;
-import ca.on.oicr.gsi.dimsum.util.reporting.reports.CaseReport;
+import ca.on.oicr.gsi.dimsum.util.reporting.ReportFormat;
+import ca.on.oicr.gsi.dimsum.util.reporting.reports.DareInputSheet;
+import ca.on.oicr.gsi.dimsum.util.reporting.reports.FullDepthSummary;
 import ca.on.oicr.gsi.dimsum.util.reporting.reports.TglTrackingReport;
 
 @RestController
@@ -34,17 +33,17 @@ public class DownloadRestController {
       @RequestBody Map<String, String> parameters, HttpServletResponse response)
       throws IOException {
 
+    ReportFormat format = Report.getFormat(parameters);
     Report report = getReport(reportName);
-    XSSFWorkbook workbook = report.writeExcelFile(caseService, parameters);
+    byte[] bytes = report.writeFile(caseService, parameters);
 
     HttpHeaders headers = new HttpHeaders();
-    headers.setContentType(new MediaType("application", "vnd.ms-excel"));
+    headers.setContentType(format.getMediaType());
     response.setHeader("Content-Disposition",
         "attachment; filename="
-            + String.format("%s-%s.xlsx", reportName, DateTimeFormatter.ISO_LOCAL_DATE.format(
-                ZonedDateTime.now())));
+            + String.format("%s-%s.%s", reportName, DateTimeFormatter.ISO_LOCAL_DATE.format(
+                ZonedDateTime.now()), format.getExtension()));
 
-    byte[] bytes = getReportBytes(workbook);
     return new HttpEntity<byte[]>(bytes, headers);
   }
 
@@ -52,17 +51,12 @@ public class DownloadRestController {
     switch (reportName) {
       case "tgl-tracking-sheet":
         return TglTrackingReport.INSTANCE;
-      case "case-report":
-        return CaseReport.INSTANCE;
+      case "full-depth-summary":
+        return FullDepthSummary.INSTANCE;
+      case "dare-input-sheet":
+        return DareInputSheet.INSTANCE;
       default:
         throw new BadRequestException("Invalid report name");
-    }
-  }
-
-  private static byte[] getReportBytes(XSSFWorkbook workbook) throws IOException {
-    try (ByteArrayOutputStream output = new ByteArrayOutputStream()) {
-      workbook.write(output);
-      return output.toByteArray();
     }
   }